Splitted HTML file into HTMLRenderer and HTMLElement.
