+++ js/Extensions/BoostCanvas.js
@@ -12,6 +12,7 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
+'use strict';
 import Chart from '../Core/Chart/Chart.js';
 import Color from '../Core/Color/Color.js';
 var color = Color.parse;
