+++ ts/Core/Renderer/SVG/SVGElementLike.d.ts
@@ -1,0 +1,31 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+/**
+ * Optional properties for the SVGElement class. Should be defined outside of
+ * this file by `declare module` pattern.
+ */
+export interface SVGElementLike {
+    // nothing here yet
+}
+
+/* *
+ *
+ *  Export
+ *
+ * */
+
+export default SVGElementLike;
